Only shows the matched user's cards
