+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-      
+        <activity android:name=".MatchingScreen"></activity>
         <activity
             android:name=".LoginActivity"
             android:label="@string/app_name"
@@ -32,7 +32,6 @@
         <activity
             android:name=".CardStack"
             android:theme="@style/AppTheme" />
-
     </application>
 
 </manifest>
