+++ app/src/main/res/layout/swipe_card.xml
@@ -21,6 +21,8 @@
         android:id="@+id/textViewCard"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:gravity="center_horizontal"
+        android:layout_marginTop="350dp"
         android:textSize="44sp"
         android:text="@string/card_name"/>
